Uses long instead of int to work with larger data

+++ SharedMediaStreamer.Domain/Interfaces/IMediaRepository.cs
@@ -4,6 +4,6 @@ namespace SharedMediaStreamer.Domain.Interfaces
 {
     public interface IMediaRepository
     {
-        MediaDetails GetMedia(int offset);
+        MediaDetails GetMedia(long offset);
     }
 }
